requirements.txt
# Original Requirements
filterpy==1.4.5
scikit-image==0.14.2
lap==0.4.0

# Requirements from YOLOv5 Repository
# Not all are used but just keeping it here for now.
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
pillow
# pycocotools>=2.0
PyYAML>=5.3
scipy>=1.4.1
tensorboard>=2.2
torch>=1.6.0
torchvision>=0.7.0
tqdm>=4.41.0